Converted the app to use Open Sourced HtmlAgilityPack instead of IronScraper, which is not.
