+++ ProductAvalibilityChecker/ProductAvalibilityChecker.csproj
@@ -3,13 +3,14 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>PAC</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="EmbedIO" Version="3.4.3" />
+    <PackageReference Include="Fizzler.Systems.HtmlAgilityPack" Version="1.2.1" />
     <PackageReference Include="HttpMultipartParser" Version="4.3.1" />
-    <PackageReference Include="IronWebScraper" Version="4.0.4.3" />
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="3.1.10" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="5.0.0" />
     <PackageReference Include="Twilio" Version="5.50.0" />
   </ItemGroup>
