+++ ProductAvalibilityChecker/WebSocketsChatServer.cs
@@ -1,0 +1,40 @@
+﻿using EmbedIO.WebSockets;
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PAC.Servers
+{
+    /// <summary>
+    /// Defines a very simple chat server.
+    /// </summary>
+    public class WebSocketsChatServer : WebSocketModule
+    {
+        public WebSocketsChatServer(string urlPath)
+            : base(urlPath, true)
+        {
+            // placeholder
+        }
+
+        /// <inheritdoc />
+        protected override Task OnMessageReceivedAsync(
+            IWebSocketContext context,
+            byte[] rxBuffer,
+            IWebSocketReceiveResult rxResult)
+            => SendToOthersAsync(context, Encoding.GetString(rxBuffer));
+
+        /// <inheritdoc />
+        protected override Task OnClientConnectedAsync(IWebSocketContext context)
+            => Task.WhenAll(
+                SendAsync(context, "Welcome to the chat room!"),
+                SendToOthersAsync(context, "Someone joined the chat room."));
+
+        /// <inheritdoc />
+        protected override Task OnClientDisconnectedAsync(IWebSocketContext context)
+            => SendToOthersAsync(context, "Someone left the chat room.");
+
+        private Task SendToOthersAsync(IWebSocketContext context, string payload)
+            => BroadcastAsync(payload, c => c != context);
+    }
+}
